Refactor HTTP request handling into HttpAdapter class
- Implement the `HttpAdapter` class to encapsulate the logic for converting Telegram requests to HTTP requests.
- Update `PsrTelegramClient` to use `HttpAdapter` for handling requests.
- Enhance unit tests to verify the functionality of `HttpAdapter`.
